Added WorkpieceCalculation interface implementation

+++ src/main/java/ru/vkarpov/dev/fbolts/service/type1d1/CostCalculation.java
@@ -4,5 +4,5 @@ package ru.vkarpov.dev.fbolts.service.type1d1;
  */
 
 public interface CostCalculation {
-    float costStudCalculation(float weightStud, float priceMetal);
+    double costStudCalculation(double weightStud, double priceMetal);
 }
